[FIX] Fix the circular dependency problem.
